Rebased Samirs achievements feature onto demo

+++ backend/controllers/courses.controller.js
@@ -1,6 +1,6 @@
-const Course = require('../models/Courses');
-const { v4: uuidv4 } = require('uuid');
-const imagekit = require('../config/imageKit');
+const Course = require("../models/Courses");
+const { v4: uuidv4 } = require("uuid");
+const imagekit = require("../config/imageKit");
 
 // Add courses
 const addCourse = async (req, res) => {
@@ -9,21 +9,18 @@ const addCourse = async (req, res) => {
     const poster = newCourse.poster;
     let newImageUrl;
     let newImageId;
-    if (newCourse.instructors && newCourse.instructors.length > 0) {
-      for (let i = 0; i < newCourse.instructors.length; i++) {
-        const instructor = newCourse.instructors[i];
-        if (instructor.imageUrl) {
-          const fileName = `image_${uuidv4()}`;
-          const uploadResponse = await imagekit.upload({
-            file: instructor.imageUrl,
-            fileName,
-            useUniqueFileName: true,
-          });
+    if (newCourse.instructor) {
+      const fileName = `image_${uuidv4()}`;
+      const uploadResponse = await imagekit.upload({
+        file: newCourse.instructor.imageUrl,
+        fileName,
+        useUniqueFileName: true,
+      });
 
-          instructor.imageUrl = uploadResponse.url;
-          instructor.imageId = uploadResponse.fileId;
-        }
-      }
+      newImageUrl = uploadResponse.url;
+      newImageId = uploadResponse.fileId;
+      newCourse.instructor.imageUrl = newImageUrl;
+      newCourse.instructor.imageId = newImageId;
     }
 
     if (poster) {
@@ -47,12 +44,12 @@ const addCourse = async (req, res) => {
     await course.save();
 
     return res.status(201).json({
-      message: 'Course added successfully',
+      message: "Course added successfully",
     });
   } catch (error) {
-    console.error('Error adding course:', error);
+    console.error("Error adding course:", error);
     return res.status(500).json({
-      message: 'Failed to add course',
+      message: "Failed to add course",
     });
   }
 };
@@ -66,7 +63,7 @@ const editCourse = async (req, res) => {
 
     if (!courseId) {
       return res.status(400).json({
-        message: 'Course ID is required',
+        message: "Course ID is required",
       });
     }
 
@@ -74,37 +71,45 @@ const editCourse = async (req, res) => {
 
     if (!course) {
       return res.status(404).json({
-        message: 'Course not found',
+        message: "Course not found",
       });
     }
 
-    if (updatedCourse.instructors && updatedCourse.instructors.length > 0) {
-      for (let i = 0; i < updatedCourse.instructors.length; i++) {
-        const instructor = updatedCourse.instructors[i];
-        const imageUrl = instructor.imageUrl;
+    if (updatedCourse.instructor) {
+      const imageUrl = updatedCourse.instructor.imageUrl;
+      if (
+        imageUrl &&
+        typeof imageUrl === "string" &&
+        !imageUrl.startsWith("http") &&
+        imageUrl !== course.instructor.imageUrl
+      ) {
+        const oldId = course.instructor?.imageId;
 
-        // Check if it's a new base64 image string
-        if (
-          imageUrl &&
-          typeof imageUrl === 'string' &&
-          !imageUrl.startsWith('http')
-        ) {
-          const fileName = `image_${uuidv4()}`;
-          const uploadResponse = await imagekit.upload({
-            file: imageUrl,
-            fileName,
-            useUniqueFileName: true,
-          });
-
-          instructor.imageUrl = uploadResponse.url;
-          instructor.imageId = uploadResponse.fileId;
+        try {
+          if (oldId) {
+            await imagekit.deleteFile(oldId);
+          }
+        } catch (error) {
+          console.log(`Failed to delete file ${oldId}: `, error.message);
         }
+
+        const fileName = `image_${uuidv4()}`;
+        const uploadResponse = await imagekit.upload({
+          file: updatedCourse.instructor.imageUrl,
+          fileName,
+          useUniqueFileName: true,
+        });
+
+        const updatedImageUrl = uploadResponse.url;
+        const updatedImageId = uploadResponse.fileId;
+        updatedCourse.instructor.imageUrl = updatedImageUrl;
+        updatedCourse.instructor.imageId = updatedImageId;
       }
     }
 
     if (
-      typeof posterUrl === 'string' &&
-      !posterUrl.startsWith('http') &&
+      typeof posterUrl === "string" &&
+      !posterUrl.startsWith("http") &&
       posterUrl !== course.posterUrl
     ) {
       const oldPosterId = course.posterId;
@@ -138,12 +143,12 @@ const editCourse = async (req, res) => {
     );
 
     return res.status(200).json({
-      message: 'Course updated successfully',
+      message: "Course updated successfully",
     });
   } catch (error) {
-    console.error('Error updating course:', error);
+    console.error("Error updating course:", error);
     return res.status(500).json({
-      message: 'Failed to update course',
+      message: "Failed to update course",
     });
   }
 };
@@ -155,7 +160,7 @@ const deleteCourse = async (req, res) => {
 
     if (!id) {
       return res.status(400).json({
-        message: 'Course ID is required',
+        message: "Course ID is required",
       });
     }
 
@@ -163,27 +168,15 @@ const deleteCourse = async (req, res) => {
 
     if (!course) {
       return res.status(404).json({
-        message: 'Course not found',
+        message: "Course not found",
       });
     }
 
     // Delete the images
-    if (course.instructors && course.instructors.length > 0) {
-      for (const instructor of course.instructors) {
-        if (instructor.imageId) {
-          try {
-            await imagekit.deleteFile(instructor.imageId);
-          } catch (error) {
-            console.log(
-              `Failed to delete file ${instructor.imageId}: `,
-              error.message
-            );
-          }
-        }
-      }
-    }
+    if (course.instructor.imageId && course.posterId) {
+      const oldId = course.instructor.imageId;
+      await imagekit.deleteFile(oldId);
 
-    if (course.posterId) {
       const oldPosterId = course.posterId;
       await imagekit.deleteFile(oldPosterId);
     }
@@ -192,12 +185,12 @@ const deleteCourse = async (req, res) => {
     await Course.findByIdAndDelete(id);
 
     return res.status(200).json({
-      message: 'Course deleted successfully',
+      message: "Course deleted successfully",
     });
   } catch (error) {
-    console.error('Error deleting course:', error);
+    console.error("Error deleting course:", error);
     return res.status(500).json({
-      message: 'Failed to delete course',
+      message: "Failed to delete course",
     });
   }
 };
@@ -208,13 +201,13 @@ const getCourses = async (req, res) => {
     const courses = await Course.find().sort({ createdAt: 1 });
 
     res.status(200).json({
-      message: 'Courses returned successfully',
+      message: "Courses returned successfully",
       coursesData: courses,
     });
   } catch (error) {
-    console.error('Error fetching courses:', error);
+    console.error("Error fetching courses:", error);
     res.status(500).json({
-      message: 'Failed to fetch courses',
+      message: "Failed to fetch courses",
     });
   }
 };
